Snapshot 2021-02-22: hello-meld-1: app loads and tests run OK. hello-meld-2: app loads OK, tests fail with redux module access problem. hello-meld-3: app loads OK, tests not yet implemented.

+++ hello-meld-1/reinstall-app.sh
@@ -1,13 +1,48 @@
 # Reinstall current MELD app using local copy of meld-clients-core
 #
 
-# For some reason, removing package-lock.json causes all sorts of warnings about obsolete packages,
-# but leaving it, the next attempt to re-build completes cleanly, despite node_modules being reset.
-#   rm package-lock.json
+MCC_NODE_MODULES="../../meld-clients-core/node_modules"
 
-rm -rf node_modules
+if [[ "$1" == "unlink" ]]; then
+    echo "Unlink meld-clients-core package..."
+    rm $(npm root)/meld-clients-core
+    return 0
+fi
+
+if [[ "$1" == "clean-only" ]]; then
+    echo "Remove built files..."
+    rm package-lock.json
+    rm -rf node_modules 
+    rm -rf lib
+    return 0
+fi
+
+if [[ "$1" == "clean" ]]; then
+    echo "Removing old package-lock.json and node_modules and lib..."
+    rm package-lock.json
+    rm -rf node_modules 
+    rm -rf lib
+    # rm $(npm root)/meld-clients-core
+fi
+
+echo "Make sure 'meld-clients-core/node_modules' is available for linking operations ..."
+if [ -d ${MCC_NODE_MODULES} ]; then
+    pass
+elif [ -d ${MCC_NODE_MODULES}_save ]; then
+    mv ${MCC_NODE_MODULES}_save ${MCC_NODE_MODULES}
+else
+	echo "Cannot find ' ${MCC_NODE_MODULES}' or '${MCC_NODE_MODULES}_save'" 
+	return 1
+fi
+
+echo "npm link meld-clients-core ..."
 npm link meld-clients-core
+echo "npm install ..."
 npm install
+echo "npm link meld-clients-core ..."
 npm link meld-clients-core
+
+# Hide `meld-clients-core/node_modules` for running app (avoid multiple react versions)
+mv ${MCC_NODE_MODULES} ${MCC_NODE_MODULES}_save
 
 # npm start
